pyproject.toml
[project]
name = "waktu-semaye"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "pytelegrambotapi>=4.31.0",
    "python-dotenv>=1.2.1",
]
